Improved obstacle and temperature zone decreassion speed

+++ SuperNautic/SuperNautic_Game/src/Core/Track/Track.hpp
@@ -60,7 +60,7 @@ private:
 	bool bInsertNormalSegment(const int index, bool testCollision);
 	void insertStructure(const int index);
 	void addWindowsAndZonesToSegment(const Segment* segment);
-	void deleteSegments(const int lengthToDelete);
+	void deleteSegments(const float lengthToDelete);
 	bool bEndTrack();
 	void placeObstacles();
 	void placeVisibilityArea();
